fix(playground): sync user preferred color scheme
